remove timeout
It works without setting the deadline. The timeout was introduced to deal
with peers who do not close the connection properly. They can exhaust all
sockets and stop normal operation of the service. However this code was
removed, because we have to confirm this problem before fixing it.

See https://github.com/DonnchaC/oniongateway/pull/3#discussion_r73783551